scaffolding(financiero_construccion_jun2026): base compartida finanzas v2
- S1: split apps/financiero/models.py -> models_base.py (monolito) + importer + stub models_finv2_mapeo.py
- S2: stub views_finv2_dashboard.DashboardFinancieroMixinV2 + inyectado como 1ra base de DashboardFinancieroView
- S3: stub apps/financiero/urls_finv2.py + registro add-only en urls.py

Habilita B1 (#120) y B2 (#122) sin tocar el monolito ni la logica existente.
Sin migraciones (la 0011 la crea B1). Sanity: py_compile verde; django runtime deferido a F4 Docker.

Refs #120 #122

🤖 Generado por F2 del skill /modulo
