+++ apps/financiero/models_finv2_mapeo.py
@@ -1,0 +1,11 @@
+"""
+Mapeo contable v2 (financiero).
+
+Stub de scaffolding plantado por F2 del skill /modulo.
+B1 (#120) llena MapeoCtaRubro aquí (modelo + migration 0011).
+"""
+from django.db import models  # noqa
+
+from apps.core.models import BaseModel  # noqa
+
+# B1 (#120) llena MapeoCtaRubro aquí
