+++ apps/financiero/urls_finv2.py
@@ -1,0 +1,8 @@
+"""
+Rutas financiero v2 (mapeo contable / carga BD contable).
+
+Stub de scaffolding plantado por F2 del skill /modulo.
+B1 (#120) agrega cargar-bd-contable / editar-mapeo.
+"""
+
+urlpatterns = []  # B1 (#120) agrega cargar-bd-contable / editar-mapeo
