+++ apps/financiero/views_finv2_dashboard.py
@@ -1,0 +1,13 @@
+"""
+Dashboard financiero v2 — mixin de indicadores técnico-financieros / ANS.
+
+Stub de scaffolding plantado por F2 del skill /modulo.
+B2 (#122) implementa aquí ``_calcular_indicadores_tecnico_financieros`` y
+``_calcular_ans`` y los engancha vía ``get_context_data``.
+"""
+
+
+class DashboardFinancieroMixinV2:
+    """Stub scaffolding. B2 (#122) implementa
+    _calcular_indicadores_tecnico_financieros / _ans."""
+    pass
